+++ processor/src/main/java/react4j/processor/ProcessorUtil.java
@@ -37,7 +37,7 @@ final class ProcessorUtil
   private static final Pattern GETTER_PATTERN = Pattern.compile( "^get([A-Z].*)$" );
   static final Pattern DEFAULT_GETTER_PATTERN = Pattern.compile( "^get([A-Z].*)Default$" );
   static final Pattern VALIDATE_PROP_PATTERN = Pattern.compile( "^validate([A-Z].*)$" );
-  static final Pattern ON_PROP_CHANGED_PATTERN = Pattern.compile( "^on([A-Z].*)Changed$" );
+  static final Pattern ON_PROP_CHANGE_PATTERN = Pattern.compile( "^on([A-Z].*)Change$" );
   private static final Pattern ISSER_PATTERN = Pattern.compile( "^is([A-Z].*)$" );
   private static final String SENTINEL_NAME = "<default>";