+++ processor/src/test/resources/bad_input/com/example/on_prop_change/ReturnOnPropChange.java
@@ -1,0 +1,27 @@
+package com.example.on_prop_change;
+
+import react4j.Component;
+import react4j.ReactNode;
+import react4j.annotations.OnPropChange;
+import react4j.annotations.Prop;
+import react4j.annotations.ReactComponent;
+
+@ReactComponent
+abstract class ReturnOnPropChange
+  extends Component
+{
+  @OnPropChange
+  int onMyPropChange( String oldValue )
+  {
+    return 0;
+  }
+
+  @Prop
+  protected abstract String getMyProp();
+
+  @Override
+  protected ReactNode render()
+  {
+    return null;
+  }
+}