+++ processor/src/test/resources/expected/com/example/on_prop_change/React4j_BooleanOnPropChange.java
@@ -1,0 +1,110 @@
+package com.example.on_prop_change;
+
+import javax.annotation.Generated;
+import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
+import jsinterop.annotations.JsConstructor;
+import jsinterop.annotations.JsPackage;
+import jsinterop.annotations.JsType;
+import jsinterop.base.Js;
+import jsinterop.base.JsPropertyMap;
+import react4j.ComponentConstructorFunction;
+import react4j.NativeAdapterComponent;
+import react4j.ReactConfig;
+
+@Generated("react4j.processor.ReactProcessor")
+class React4j_BooleanOnPropChange extends BooleanOnPropChange {
+  @Nonnull
+  private static ComponentConstructorFunction getConstructorFunction() {
+    final ComponentConstructorFunction componentConstructor = ( ReactConfig.shouldStoreDebugDataAsState() || ReactConfig.shouldValidatePropValues() ) ? NativeReactComponent::new : LiteNativeReactComponent::new;
+    if ( ReactConfig.enableComponentNames() ) {
+      Js.asPropertyMap( componentConstructor ).set( "displayName", "BooleanOnPropChange" );
+    }
+    return componentConstructor;
+  }
+
+  @Override
+  protected boolean getMyProp() {
+    return props().getAny( Props.myProp ).asBoolean();
+  }
+
+  @Override
+  protected boolean reportPropChanges(@Nonnull final JsPropertyMap<Object> props,
+      @Nonnull final JsPropertyMap<Object> nextProps, final boolean inComponentDidUpdate) {
+    boolean modified = false;
+    if ( !Js.isTripleEqual( props.get( Props.myProp ), nextProps.get( Props.myProp ) ) ) {
+      if ( inComponentDidUpdate ) {
+        onMyPropChange( props.getAny( Props.myProp ).asBoolean() );
+      }
+      modified = true;
+    }
+    return modified;
+  }
+
+  static final class Factory {
+    static final ComponentConstructorFunction TYPE = getConstructorFunction();
+  }
+
+  static final class Props {
+    static final String myProp = ReactConfig.shouldMinimizePropKeys() ? "a" : "myProp";
+  }
+
+  @JsType(
+      isNative = true,
+      namespace = JsPackage.GLOBAL,
+      name = "?"
+  )
+  interface LiteLifecycle {
+    void componentDidUpdate(@Nonnull JsPropertyMap<Object> prevProps);
+  }
+
+  @JsType(
+      isNative = true,
+      namespace = JsPackage.GLOBAL,
+      name = "?"
+  )
+  interface Lifecycle {
+    void componentDidMount();
+
+    void componentDidUpdate(@Nonnull JsPropertyMap<Object> prevProps);
+  }
+
+  private static final class LiteNativeReactComponent extends NativeAdapterComponent<BooleanOnPropChange> implements LiteLifecycle {
+    @JsConstructor
+    LiteNativeReactComponent(@Nullable final JsPropertyMap<Object> props) {
+      super( props );
+    }
+
+    @Override
+    protected BooleanOnPropChange createComponent() {
+      return new React4j_BooleanOnPropChange();
+    }
+
+    @Override
+    public void componentDidUpdate(@Nonnull final JsPropertyMap<Object> prevProps) {
+      performComponentDidUpdate( prevProps );
+    }
+  }
+
+  private static final class NativeReactComponent extends NativeAdapterComponent<BooleanOnPropChange> implements Lifecycle {
+    @JsConstructor
+    NativeReactComponent(@Nullable final JsPropertyMap<Object> props) {
+      super( props );
+    }
+
+    @Override
+    protected BooleanOnPropChange createComponent() {
+      return new React4j_BooleanOnPropChange();
+    }
+
+    @Override
+    public void componentDidMount() {
+      performComponentDidMount();
+    }
+
+    @Override
+    public void componentDidUpdate(@Nonnull final JsPropertyMap<Object> prevProps) {
+      performComponentDidUpdate( prevProps );
+    }
+  }
+}