+++ processor/src/test/resources/input/com/example/on_prop_change/BooleanOnPropChange.java
@@ -1,0 +1,26 @@
+package com.example.on_prop_change;
+
+import react4j.Component;
+import react4j.ReactNode;
+import react4j.annotations.OnPropChange;
+import react4j.annotations.Prop;
+import react4j.annotations.ReactComponent;
+
+@ReactComponent
+abstract class BooleanOnPropChange
+  extends Component
+{
+  @OnPropChange
+  void onMyPropChange( boolean oldValue )
+  {
+  }
+
+  @Prop
+  protected abstract boolean getMyProp();
+
+  @Override
+  protected ReactNode render()
+  {
+    return null;
+  }
+}