+++ processor/src/test/resources/input/com/example/on_prop_change/NullableOnPropChange.java
@@ -1,0 +1,28 @@
+package com.example.on_prop_change;
+
+import javax.annotation.Nullable;
+import react4j.Component;
+import react4j.ReactNode;
+import react4j.annotations.OnPropChange;
+import react4j.annotations.Prop;
+import react4j.annotations.ReactComponent;
+
+@ReactComponent
+abstract class NullableOnPropChange
+  extends Component
+{
+  @OnPropChange
+  void onMyPropChange( @Nullable String oldValue )
+  {
+  }
+
+  @Prop
+  @Nullable
+  protected abstract String getMyProp();
+
+  @Override
+  protected ReactNode render()
+  {
+    return null;
+  }
+}